refactor(cache): separate module build cache from backend
