+++ crates/rspack_core/src/compilation/build_module_graph/graph_updater/repair/context.rs
@@ -9,9 +9,8 @@ use crate::{
   Compilation, CompilationId, CompilerId, CompilerOptions, CompilerPlatform, DependencyTemplate,
   DependencyTemplateType, DependencyType, ExportsInfoArtifact, FileSystemInfo, ModuleFactory,
   ResolverFactory, RuntimeTemplate, SharedPluginDriver, ValueCacheVersions,
-  incremental::Incremental,
-  module_graph::ModuleGraph,
-  new_cache::{Cache, CacheFacade},
+  compilation::build_module_graph::module_build_cache::ModuleBuildCache, incremental::Incremental,
+  module_graph::ModuleGraph, new_cache::Cache,
 };
 
 #[derive(Debug)]
@@ -33,7 +32,7 @@ pub struct TaskContext {
   pub dependency_templates: HashMap<DependencyTemplateType, Arc<dyn DependencyTemplate>>,
   pub runtime_template: RuntimeTemplate,
   pub(crate) cache: Cache,
-  pub(crate) module_cache: Option<CacheFacade>,
+  pub(crate) module_build_cache: Option<ModuleBuildCache>,
   pub value_cache_versions: ValueCacheVersions,
 
   pub artifact: BuildModuleGraphArtifact,
@@ -62,7 +61,7 @@ impl TaskContext {
       intermediate_fs: compilation.intermediate_filesystem.clone(),
       output_fs: compilation.output_filesystem.clone(),
       runtime_template: RuntimeTemplate::new(compilation.options.clone()),
-      module_cache: compilation.module_cache.clone(),
+      module_build_cache: compilation.module_build_cache.clone(),
       cache: compilation.cache.clone(),
       value_cache_versions: compilation.value_cache_versions.clone(),
       artifact,
@@ -103,7 +102,7 @@ impl TaskContext {
       false,
       compiler_context,
     );
-    compilation.module_cache = None;
+    compilation.module_build_cache = None;
     compilation.runtime_template =
       RuntimeTemplate::for_module_execution(self.compiler_options.clone());
     compilation.dependency_factories = self.dependency_factories.clone();
