+++ crates/rspack_core/src/compilation/build_module_graph/mod.rs
@@ -1,5 +1,6 @@
 mod graph_updater;
 mod lazy_barrel_artifact;
+pub(crate) mod module_build_cache;
 mod module_executor;
 pub mod pass;
